Remove 'init'
This change removes the init call and all work now happens in the
generate phase.

Init is always followed by generate when ArgoCD calls us, so this was
an extra complication which doesn't give us any benefit and makes
implementing pre-processor plugins harder.

'init' as a CLI parameter is left in doing no work, but the
recommendation is to remove it from your argocd configmap so we never
get called with init. A future release should remove this backwards
compatibility fix.